+++ frontend/src/components/createDrink/IngredientPicker.js
@@ -1,14 +1,29 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { View } from 'react-native';
-import { TextInput } from 'react-native-paper';
+import { Button, TextInput } from 'react-native-paper';
 import RNPickerSelect from 'react-native-picker-select';
 
 export default function IngredientPicker(props) {
+  const [ingredient, setIngredient] = useState({});
+  const [quantity, setQuantity] = useState(0);
+  const [units, setUnits] = useState('');
+
+  const ingredientValueChange = (id, i) => {
+    if (id) {
+      setIngredient({ ingredient: id, name: props.ingredientsArray[i - 1].label });
+    }
+  };
+
+  const quantityValueChange = (newQuantity) => setQuantity(newQuantity);
+
+  const unitValueChange = (newUnits) => setUnits(newUnits);
+
   return (
     <View>
-      <RNPickerSelect placeholder={{ label: 'Select an ingredient...' }} onValueChange={props.ingredientValueChange} items={props.ingredientsArray} />
-      <TextInput keyboardType="decimal-pad" onChangeText={props.quantityValueChange} />
-      <RNPickerSelect placeholder={{ label: 'Select a measurement unit...' }} onValueChange={props.unitValueChange} items={props.pickerUnits} />
+      <RNPickerSelect placeholder={{ label: 'Select an ingredient...' }} onValueChange={ingredientValueChange} items={props.ingredientsArray} />
+      <TextInput keyboardType="decimal-pad" onChangeText={quantityValueChange} />
+      <RNPickerSelect placeholder={{ label: 'Select a measurement unit...' }} onValueChange={unitValueChange} items={props.pickerUnits} />
+      <Button onPress={props.addIngredient.bind(this, { ingredient, quantity, units })}>Add</Button>
     </View>
   );
 }
